pyproject.toml
[tool.poetry]
name = "tut_05_rest_framework"
version = "0.1.0"
description = ""
authors = ["Roman-Ojha <razzroman98@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
PyYAML = "^6.0"
requests = "^2.28.1"
django-cors-headers = "^3.13.0"
autopep8 = "^1.6.0"
algoliasearch-django = "2.0"
python-decouple = "^3.6"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
